apps/concierge-svc/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.30.0
sqlmodel==0.0.19
aiosqlite==0.20.0
httpx==0.27.0
aiokafka==0.10.0
redis==5.0.4
pydantic-settings==2.2.1
orjson==3.10.6
tenacity==8.3.0
python-dotenv==1.0.1
numpy==1.26.4
pytest==8.2.1